Move entity definitions into their own file
