+++ packages/@blockprotocol/graph/src/types.ts
@@ -1,91 +1,17 @@
-import { MessageCallback, UnknownRecord } from "@blockprotocol/core";
+import { MessageCallback } from "@blockprotocol/core";
 
-// ----------------------------- ENTITIES ----------------------------- //
+import {
+  AggregateEntitiesData,
+  AggregateEntitiesResult,
+  AggregateOperationInput,
+  CreateEntityData,
+  DeleteEntityData,
+  Entity,
+  GetEntityData,
+  UpdateEntityData,
+} from "./types/entity";
 
-export type Entity<
-  Properties extends Record<string, unknown> | null = Record<string, unknown>,
-> = {
-  entityId: string;
-  entityTypeId?: string;
-} & (Properties extends null ? {} : { properties: Properties });
-
-export type CreateEntityData = {
-  entityTypeId: string;
-  properties: UnknownRecord;
-  links?: Omit<
-    CreateLinkData,
-    "sourceAccountId" | "sourceEntityId" | "sourceEntityTypeId"
-  >[];
-};
-
-export type GetEntityData = {
-  entityId: string;
-};
-
-export type UpdateEntityData = {
-  entityId: string;
-  properties: UnknownRecord;
-};
-
-export type DeleteEntityData = {
-  entityId: string;
-};
-
-export type FilterOperatorType =
-  | FilterOperatorRequiringValue
-  | FilterOperatorWithoutValue;
-
-export type FilterOperatorWithoutValue = "IS_EMPTY" | "IS_NOT_EMPTY";
-
-export type FilterOperatorRequiringValue =
-  | "CONTAINS"
-  | "DOES_NOT_CONTAIN"
-  | "IS"
-  | "IS_NOT"
-  | "STARTS_WITH"
-  | "ENDS_WITH";
-
-export type MultiFilterOperatorType = "AND" | "OR";
-
-export type MultiFilter = {
-  filters: (
-    | {
-        field: string;
-        operator: FilterOperatorRequiringValue;
-        value: string;
-      }
-    | { field: string; operator: FilterOperatorWithoutValue }
-  )[];
-  operator: MultiFilterOperatorType;
-};
-
-export type Sort = {
-  field: string;
-  desc?: boolean | undefined | null;
-};
-
-export type MultiSort = Sort[];
-
-export type AggregateOperationInput = {
-  entityTypeId?: string | null;
-  pageNumber?: number | null;
-  itemsPerPage?: number | null;
-  multiSort?: MultiSort | null;
-  multiFilter?: MultiFilter | null;
-};
-
-export type AggregateEntitiesData = {
-  operation: AggregateOperationInput;
-};
-
-export type AggregateEntitiesResult<T extends Entity | EntityType> = {
-  results: T[];
-  operation: AggregateOperationInput &
-    Required<Pick<AggregateOperationInput, "pageNumber" | "itemsPerPage">> & {
-      pageCount?: number | null;
-      totalCount?: number | null;
-    };
-};
+export * from "./types/entity";
 
 // ------------------------ OTHER FUNCTIONS --------------------------- //
