+++ packages/@blockprotocol/graph/src/types/entity.ts
@@ -1,0 +1,88 @@
+import { UnknownRecord } from "@blockprotocol/core";
+
+import { CreateLinkData, EntityType } from "../types";
+
+export type Entity<
+  Properties extends Record<string, unknown> | null = Record<string, unknown>,
+> = {
+  entityId: string;
+  entityTypeId?: string;
+} & (Properties extends null ? {} : { properties: Properties });
+
+export type CreateEntityData = {
+  entityTypeId: string;
+  properties: UnknownRecord;
+  links?: Omit<
+    CreateLinkData,
+    "sourceAccountId" | "sourceEntityId" | "sourceEntityTypeId"
+  >[];
+};
+
+export type GetEntityData = {
+  entityId: string;
+};
+
+export type UpdateEntityData = {
+  entityId: string;
+  properties: UnknownRecord;
+};
+
+export type DeleteEntityData = {
+  entityId: string;
+};
+
+export type FilterOperatorType =
+  | FilterOperatorRequiringValue
+  | FilterOperatorWithoutValue;
+
+export type FilterOperatorWithoutValue = "IS_EMPTY" | "IS_NOT_EMPTY";
+
+export type FilterOperatorRequiringValue =
+  | "CONTAINS"
+  | "DOES_NOT_CONTAIN"
+  | "IS"
+  | "IS_NOT"
+  | "STARTS_WITH"
+  | "ENDS_WITH";
+
+export type MultiFilterOperatorType = "AND" | "OR";
+
+export type MultiFilter = {
+  filters: (
+    | {
+        field: string;
+        operator: FilterOperatorRequiringValue;
+        value: string;
+      }
+    | { field: string; operator: FilterOperatorWithoutValue }
+  )[];
+  operator: MultiFilterOperatorType;
+};
+
+export type Sort = {
+  field: string;
+  desc?: boolean | undefined | null;
+};
+
+export type MultiSort = Sort[];
+
+export type AggregateOperationInput = {
+  entityTypeId?: string | null;
+  pageNumber?: number | null;
+  itemsPerPage?: number | null;
+  multiSort?: MultiSort | null;
+  multiFilter?: MultiFilter | null;
+};
+
+export type AggregateEntitiesData = {
+  operation: AggregateOperationInput;
+};
+
+export type AggregateEntitiesResult<T extends Entity | EntityType> = {
+  results: T[];
+  operation: AggregateOperationInput &
+    Required<Pick<AggregateOperationInput, "pageNumber" | "itemsPerPage">> & {
+      pageCount?: number | null;
+      totalCount?: number | null;
+    };
+};
